splitting rsvp tests into multiple files
